+++ Source/Data.swift
@@ -233,9 +233,9 @@ public struct CardDetails {
     public let cardNetwork: CardNetwork?
     
     static func fromDictionary(dict: JSONDictionary) -> CardDetails {
-        let lastFour = dict["cardLastFour"] as! String
-        let endDate = dict["endDate"] as! String
-        let cardToken = dict["cardToken"] as! String
+        let lastFour = dict["cardLastfour"] as? String
+        let endDate = dict["endDate"] as? String
+        let cardToken = dict["cardToken"] as? String
         // TODO: parse dict["cardType"] into CardNetwork
         return CardDetails(cardLastFour: lastFour, endDate: endDate, cardToken: cardToken, cardNetwork: nil)
     }
@@ -266,115 +266,6 @@ let ISO8601DateFormatter: NSDateFormatter = {
     return dateFormatter
 }()
 
-
-/**
-*  TransactionResult will hold all the information from a transaction response
-*/
-public struct TransactionData {
-    /// our reference for this transaction. Keep track of this as it's needed to process refunds or collections later
-    public let receiptID: String
-    /// Your original reference for this payment
-    public let yourPaymentReference: String
-    /// The type of Transaction, either "Payment" or "Refund"
-    public let type: TransactionType
-    /// date and time of the Transaction including time zone offset
-    public let createdAt: NSDate
-    /// The result of this transactions, this will either be "Success" or "Declined"
-    public let result: TransactionResult
-    /// A message detailing the result.
-    public let message: String
-    /// The number (e.g. "123-456" or "654321") identifying the Merchant to whom payment has been made
-    public let judoID: String
-    /// The trading name of the Merchant to whom payment has been made
-    public let merchantName: String
-    /// How the Merchant will appear on the Consumers statement
-    public let appearsOnStatementAs: String
-    /// If present this will show the total value of refunds made against the original payment
-    public let refunds: Amount?
-    /// This is the original value of this transaction before refunds
-    public let originalAmount: Amount?
-    /// This will show the remaining balance of the transaction after refunds. You cannot refund more than the original payment
-    public let netAmount: Amount
-    /// This is the value of this transaction (if refunds available it is the amount after refunds)
-    public let amount: Amount
-    /// Information about the card used in this transaction
-    public let cardDetails: CardDetails
-    /// details of the Consumer for use in repeat payments
-    public let consumer: Consumer
-    
-    
-    /**
-    create a TransactionData Object from a dictionary
-    
-    - Parameter dict: the dictionary
-    
-    - Returns: a TransactionData object
-    */
-    static func fromDictionary(dict: JSONDictionary) throws -> TransactionData {
-        guard let receiptID = dict["receiptId"] as? String else { throw JudoError.ResponseParseError }
-        
-        guard let yourPaymentReference = dict["yourPaymentReference"] as? String else { throw JudoError.ResponseParseError }
-        
-        guard let typeString = dict["type"] as? String else { throw JudoError.ResponseParseError }
-        guard let type = TransactionType(rawValue: typeString) else { throw JudoError.ResponseParseError }
-        
-        guard let createdAtString = dict["createdAt"] as? String else { throw JudoError.ResponseParseError }
-        guard let createdAt = ISO8601DateFormatter.dateFromString(createdAtString) else { throw JudoError.ResponseParseError }
-        
-        guard let resultString = dict["result"] as? String else { throw JudoError.ResponseParseError }
-        guard let result = TransactionResult(rawValue: resultString) else { throw JudoError.ResponseParseError }
-        
-        guard let message = dict["message"] as? String else { throw JudoError.ResponseParseError }
-        
-        guard let judoID = dict["judoID"] as? String else { throw JudoError.ResponseParseError }
-        
-        guard let merchantName = dict["merchantName"] as? String else { throw JudoError.ResponseParseError }
-        
-        guard let appearsOnStatementAs = dict["appearsOnStatementAs"] as? String else { throw JudoError.ResponseParseError }
-        
-        let refundsDouble = dict["refunds"] as? Double
-        let refunds = Amount(refundsDouble)
-
-        let originalAmountDouble = dict["originalAmount"] as? Double
-        let originalAmount = Amount(originalAmountDouble)
-        
-        guard let netAmountDouble = dict["netAmount"] as? Double else { throw JudoError.ResponseParseError }
-        let netAmount = Amount(netAmountDouble)
-        
-        guard let amountDouble = dict["amount"] as? Double else { throw JudoError.ResponseParseError }
-        let amount = Amount(amountDouble)
-        
-        guard let cardDetailsDict = dict["cardDetails"] as? JSONDictionary else { throw JudoError.ResponseParseError }
-        let cardDetails = CardDetails.fromDictionary(cardDetailsDict)
-        
-        guard let consumerDict = dict["consumer"] as? JSONDictionary else { throw JudoError.ResponseParseError }
-        let consumer = Consumer.fromDictionary(consumerDict)
-        
-        return TransactionData(receiptID: receiptID, yourPaymentReference: yourPaymentReference, type: type, createdAt: createdAt, result: result, message: message, judoID: judoID, merchantName: merchantName, appearsOnStatementAs: appearsOnStatementAs, refunds: refunds, originalAmount: originalAmount, netAmount: netAmount, amount: amount, cardDetails: cardDetails, consumer: consumer)
-    }
-}
-
-
-/**
-Type of Transaction
-
-- Payment: a Payment Transaction
-- Refund:  a Refund Transaction
-*/
-public enum TransactionType: String {
-    case Payment="Payment", Refund="Refund"
-}
-
-
-/**
-Result of a Transaction
-
-- Success:  successful transaction
-- Declined: declined transaction
-*/
-public enum TransactionResult: String {
-    case Success="Success", Declined="Declined"
-}
 
 
 // MARK: Pagination