+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 All notable changes to this project will be documented in this file.
 'Judo' adheres to [Semantic Versioning](http://semver.org/).
 
-- `0.1.x` Releases - [0.1.0](#010) | [0.1.3](#013)
+- `0.1.x` Releases - [0.1.0](#010) | [0.1.3](#013) | [0.1.5](#015)
+
+---
+## [0.1.5](https://github.com/JudoPay/Judo-Swift/tag/0.1.5)
+Released on 2015-08-03
+
+#### Added
+- Types as responses instead of JSON Dictionaries
+- Pagination for Array type responses
+	- Updated by [Hamon Ben Riazy](https://github.com/ryce).
 
 ---
 ## [0.1.3](https://bitbucket.org/judo/judo-swift/commits/tag/0.1.3)